Refactor: 파일 구조 개선 및 폴더 재구성 #40
