+++ app/src/main/java/com/example/hrr_android/access/model/LoginResponse.kt
@@ -1,4 +1,4 @@
-package com.example.hrr_android
+package com.example.hrr_android.access.model
 
 data class LoginResponse(
     val accessToken: String,
